backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
greenlet==3.0.3
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.12
email-validator==2.2.0
aiofiles==24.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pymysql==1.1.1
cryptography==43.0.1
